pow: Allow negative exponents
Supporting negative exponents is fairly straightforward, we
just need to be careful to start with enough precision when doing
the initial invertion.

+++ src/lib.rs
@@ -746,14 +746,14 @@ impl BigDecimal {
     /// assert_eq!(n.pow(3000000000), "9.816204233623505350831385407878283564899139328691307267002649220552261820356883420275966921502700387e903089986".parse().unwrap());
     /// ```
     #[inline]
-    pub fn pow(&self, exp: u32) -> BigDecimal {
+    pub fn pow(&self, exp: i64) -> BigDecimal {
         self.pow_with_context(exp, &Context::default())
     }
 
     /// Raises the number to an integer power, using context for precision and rounding
     ///
     #[inline]
-    pub fn pow_with_context(&self, exp: u32, ctx: &Context) -> BigDecimal {
+    pub fn pow_with_context(&self, exp: i64, ctx: &Context) -> BigDecimal {
         if self.is_zero() || self.is_one() {
             return self.clone();
         }
